Fixed flycheck errors (I think) for real this time.
Not sure if this is the best way to go about it. Followed
a Reddit comment from someone having a similar issue
(suggested wrapping the bootstrap in eval-and-compile)
and then needed to add :defines and :functions for everything
used by use-package.

Will try to quickly check if uncommenting the flycheck inherit
would also fix the issue. (Yeah, I should've checked that first...)